Extract configuration and builder to config.go
