+++ cpu.go
@@ -27,41 +27,7 @@ package cpu6502
 
 import (
 	"fmt"
-	"log"
 )
-
-// CPUConfig holds configuration options for CPU creation
-type CPUConfig struct {
-	// Variant specifies the CPU variant (NMOS, CMOS, Ricoh)
-	Variant CPUVariant
-
-	// ErrorHandler defines how errors are handled
-	ErrorHandler ErrorHandler
-
-	// StrictMode halts execution on illegal opcodes
-	StrictMode bool
-
-	// EnableDecimalMode allows disabling decimal mode even on variants that support it
-	EnableDecimalMode bool
-
-	// EnableInstructionCache enables instruction caching for performance
-	EnableInstructionCache bool
-
-	// InstructionCacheSize sets the cache size (default: 256 entries)
-	InstructionCacheSize int
-}
-
-// DefaultConfig returns a configuration with sensible defaults
-func DefaultConfig() CPUConfig {
-	return CPUConfig{
-		Variant:                VariantNMOS6502,
-		ErrorHandler:           &LoggingErrorHandler{Logger: log.Default()},
-		StrictMode:             false,
-		EnableDecimalMode:      true,
-		EnableInstructionCache: true,
-		InstructionCacheSize:   256,
-	}
-}
 
 // CPU represents a MOS Technology 6502 microprocessor.
 //
@@ -128,133 +94,6 @@ type CPU struct {
 
 	// Performance optimization
 	instrCache *InstructionCache
-}
-
-// NewCPU creates a new 6502 CPU instance with default configuration.
-//
-// The CPU is initialized with:
-//   - All registers cleared
-//   - Stack pointer at $FD
-//   - Status flags: U and I set
-//   - NMOS 6502 variant
-//   - Logging error handler
-//
-// The CPU must be reset before execution:
-//
-//	cpu := NewCPU(bus)
-//	cpu.Reset() // Loads PC from reset vector at $FFFC/FD
-//
-// For custom configuration, use NewCPUWithConfig instead.
-func NewCPU(bus Bus) *CPU {
-	return NewCPUWithConfig(bus, DefaultConfig())
-}
-
-// NewCPUWithVariant creates a new CPU with specified variant
-func NewCPUWithVariant(bus Bus, variant CPUVariant) *CPU {
-	config := DefaultConfig()
-	config.Variant = variant
-	return NewCPUWithConfig(bus, config)
-}
-
-// NewCPUWithErrorHandler creates a new CPU with default NMOS 6502 variant and custom error handler
-func NewCPUWithErrorHandler(bus Bus, handler ErrorHandler) *CPU {
-	config := DefaultConfig()
-	config.ErrorHandler = handler
-	return NewCPUWithConfig(bus, config)
-}
-
-// NewCPUWithVariantAndErrorHandler creates a new CPU with specified variant and error handler
-func NewCPUWithVariantAndErrorHandler(bus Bus, variant CPUVariant, handler ErrorHandler) *CPU {
-	config := DefaultConfig()
-	config.Variant = variant
-	config.ErrorHandler = handler
-	return NewCPUWithConfig(bus, config)
-}
-
-// NewCPUWithConfig creates a new CPU with full configuration
-func NewCPUWithConfig(bus Bus, config CPUConfig) *CPU {
-	// Apply strict mode to error handler if requested
-	errorHandler := config.ErrorHandler
-	if config.StrictMode {
-		errorHandler = &StrictErrorHandler{}
-	}
-
-	c := &CPU{
-		bus:          bus,
-		P:            U | I,
-		SP:           0xFD,
-		variant:      config.Variant,
-		errorHandler: errorHandler,
-	}
-
-	// Initialize instruction cache if enabled
-	if config.EnableInstructionCache {
-		c.instrCache = NewInstructionCache()
-	}
-
-	c.buildLookupTable()
-
-	// Apply decimal mode configuration
-	if !config.EnableDecimalMode {
-		c.P &^= D // Clear decimal flag
-	}
-
-	return c
-}
-
-// CPUBuilder provides a fluent interface for CPU configuration
-type CPUBuilder struct {
-	bus    Bus
-	config CPUConfig
-}
-
-// NewBuilder creates a new CPU builder
-func NewBuilder(bus Bus) *CPUBuilder {
-	return &CPUBuilder{
-		bus:    bus,
-		config: DefaultConfig(),
-	}
-}
-
-// WithVariant sets the CPU variant
-func (b *CPUBuilder) WithVariant(variant CPUVariant) *CPUBuilder {
-	b.config.Variant = variant
-	return b
-}
-
-// WithStrictMode enables strict mode
-func (b *CPUBuilder) WithStrictMode() *CPUBuilder {
-	b.config.StrictMode = true
-	return b
-}
-
-// WithErrorHandler sets a custom error handler
-func (b *CPUBuilder) WithErrorHandler(handler ErrorHandler) *CPUBuilder {
-	b.config.ErrorHandler = handler
-	return b
-}
-
-// DisableDecimalMode disables decimal mode
-func (b *CPUBuilder) DisableDecimalMode() *CPUBuilder {
-	b.config.EnableDecimalMode = false
-	return b
-}
-
-// DisableInstructionCache disables the instruction cache
-func (b *CPUBuilder) DisableInstructionCache() *CPUBuilder {
-	b.config.EnableInstructionCache = false
-	return b
-}
-
-// WithInstructionCacheSize sets the instruction cache size
-func (b *CPUBuilder) WithInstructionCacheSize(size int) *CPUBuilder {
-	b.config.InstructionCacheSize = size
-	return b
-}
-
-// Build creates the configured CPU
-func (b *CPUBuilder) Build() *CPU {
-	return NewCPUWithConfig(b.bus, b.config)
 }
 
 // Variant returns the CPU variant
